Xcode 13.4.1をサポート

+++ CoreTextSample/Sample12.m
@@ -71,7 +71,7 @@
         // 行のベースラインpointを取得
         CGPoint origin = *(origins + i);
         // 行からdescentを取得
-        float   descent;
+        CGFloat   descent;
         CTLineGetTypographicBounds(line, nil, &descent, nil);
         // originはベースラインなのでdescentを引けば行の原点(その行の左上端)になる。
         CGContextSetTextPosition(context, origin.x, origin.y - descent);
